# Remove Docker from Travis CI, they're run on GHA