For https://github.com/python-pillow/Pillow/issues/3606.

Changes proposed in this pull request:

 * PR https://github.com/python-pillow/Pillow/pull/4316 added coverage to Docker on the faster GitHub Actions
 * We can now remove Docker builds from Travis CI


# Before

* Ran for 13 min 5 sec
* Total time 50 min 59 sec

![image](https://user-images.githubusercontent.com/1324225/71777873-e457dd80-2fae-11ea-8f2f-1a806327bf48.png)


# After

* Ran for 9 min 9 sec
* Total time 19 min 45 sec

![image](https://user-images.githubusercontent.com/1324225/71777852-a955aa00-2fae-11ea-8270-8f6d4a66771f.png)